src/requirements.txt
astroid==2.3.3
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
future==0.18.2
google-api-core==1.14.3
google-auth==1.8.2
google-cloud-core==1.1.0
google-cloud-datastore==1.10.0
google-cloud-logging==1.14.0
google-cloud-ndb==0.2.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
lxml==4.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pylint==2.4.4
python-dotenv==0.10.3
python-telegram-bot==12.2.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
rsa==4.0
six==1.13.0
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
